requirements.txt
Django>=3.0,<4.0
django-mailbox==4.8.2
django-picklefield==3.0.1
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
google-api-python-client==1.12.8
google-auth-oauthlib==0.4.2
pandas==1.2.1
psycopg2-binary>=2.8
scikit-learn==0.24.1
social-auth-app-django==4.0
